add intermediate abstraction layer to separate incremental and adaptive behavior
