+++ incremental/src/main/java/net/automatalib/incremental/mealy/AbstractMealyBuilder.java
@@ -26,15 +26,14 @@ import net.automatalib.visualization.VisualizationHelper;
 import net.automatalib.words.Word;
 import net.automatalib.words.WordBuilder;
 
-public abstract class AbstractIncrementalMealyBuilder<I, O> {
-
-    public abstract boolean lookup(Word<? extends I> inputWord, List<? super O> output);
+public abstract class AbstractMealyBuilder<I, O> implements MealyBuilder<I, O> {
 
     public boolean hasDefinitiveInformation(Word<? extends I> word) {
         List<O> unused = new ArrayList<>(word.length());
         return lookup(word, unused);
     }
 
+    @Override
     public Word<O> lookup(Word<? extends I> inputWord) {
         WordBuilder<O> wb = new WordBuilder<>(inputWord.size());
         lookup(inputWord, wb);
